lab_1/requirements.txt
biopython
numpy
scipy